chore:renamed screen groups to stack
