+++ src/frontend/Navigation/AppStack.tsx
@@ -2,11 +2,11 @@ import React from 'react';
 import {createNativeStackNavigator} from '@react-navigation/native-stack';
 import {NativeStackNavigationOptions} from '@react-navigation/native-stack/lib/typescript/src/types';
 import {WHITE} from '../lib/styles';
-import {AppList} from './ScreenGroups/AppScreens';
+import {AppList} from './Stack/AppScreens';
 import {CustomHeaderLeft} from '../sharedComponents/CustomHeaderLeft';
-import {DeviceNamingSceens} from './ScreenGroups/DeviceNamingScreens';
+import {DeviceNamingSceens} from './Stack/DeviceNamingScreens';
 
-export {createDefaultScreenGroup} from './ScreenGroups/AppScreens';
+export {createDefaultScreenGroup} from './Stack/AppScreens';
 
 // export type AppStackList = AppList & OnboardingStackList & IccaStackList;
 export type AppStackList = AppList & DeviceNamingSceens;
