+++ src/frontend/Navigation/Stack/TabBar/CameraTabBarIcon.tsx
@@ -1,6 +1,6 @@
 import React, {FC} from 'react';
-import {TabBarIconProps} from '../../types';
-import {TabBarIcon} from './TabBarIcon';
+import {TabBarIconProps} from '../../types.ts';
+import {TabBarIcon} from './TabBarIcon.tsx';
 import {useTabNavigationStore} from '../../../hooks/useTabNavigationStore.ts';
 
 export const CameraTabBarIcon: FC<TabBarIconProps> = props => {
