+++ src/frontend/Navigation/Stack/TabBar/TabBarIcon.tsx
