+++ src/frontend/hooks/useNavigationWithTypes.ts
@@ -2,7 +2,7 @@ import {BottomTabNavigationProp} from '@react-navigation/bottom-tabs';
 import {CompositeNavigationProp, useNavigation} from '@react-navigation/native';
 import {NativeStackNavigationProp} from '@react-navigation/native-stack';
 import {AppStackList} from '../Navigation/AppStack';
-import {HomeTabsList} from '../Navigation/ScreenGroups/AppScreens';
+import {HomeTabsList} from '../Navigation/Stack/AppScreens';
 
 export const useNavigationFromRoot = () =>
   useNavigation<NativeStackNavigationProp<AppStackList>>();
