+++ src/frontend/screens/IntroToCoMapeo.tsx
@@ -8,7 +8,7 @@ import {defineMessages, useIntl} from 'react-intl';
 import {Text} from '../sharedComponents/Text';
 import {Button} from '../sharedComponents/Button';
 import {NativeStackScreenProps} from '@react-navigation/native-stack';
-import {DeviceNamingSceens} from '../Navigation/ScreenGroups/DeviceNamingScreens';
+import {DeviceNamingSceens} from '../Navigation/Stack/DeviceNamingScreens';
 
 const m = defineMessages({
   isNow: {
